VisualiserTool : Move header to `private`
