+++ include/GafferSceneUI/Private/VisualiserTool.h
