+++ src/GafferSceneUI/VisualiserTool.cpp
@@ -34,7 +34,7 @@
 //
 //////////////////////////////////////////////////////////////////////////
 
-#include "GafferSceneUI/VisualiserTool.h"
+#include "GafferSceneUI/Private/VisualiserTool.h"
 
 #include "GafferSceneUI/SceneView.h"
 #include "GafferSceneUI/ScriptNodeAlgo.h"
